Remove warning suppressions and fix many casting bugs
This change removes most warning suppressions, in particular
from production code. In the past we inherit a bunch of -Wno-xxx
suppressions required to build gtest and libprotobuf headers.
Doing so, however, caused the suppressions to propagate back to
the translation units that were including any protobuf header
or any auto-generated .pb.h stub.
This change moves the gtest and probobuf header to be a
system include (-isystem vs -I). Doing so implicitly blacklists
any compiler warning on the headers.
This CL then re-enables warnings and deals with the fall out of
fixes that came out of this.

Bug: 77316877
Test: pefetto_unittests / perfetto_integrationtests

+++ include/perfetto/tracing/core/inode_file_config.h
@@ -92,14 +92,14 @@ class PERFETTO_EXPORT InodeFileConfig {
   void FromProto(const perfetto::protos::InodeFileConfig&);
   void ToProto(perfetto::protos::InodeFileConfig*) const;
 
-  uint64_t scan_interval_ms() const { return scan_interval_ms_; }
-  void set_scan_interval_ms(uint64_t value) { scan_interval_ms_ = value; }
+  uint32_t scan_interval_ms() const { return scan_interval_ms_; }
+  void set_scan_interval_ms(uint32_t value) { scan_interval_ms_ = value; }
 
-  uint64_t scan_delay_ms() const { return scan_delay_ms_; }
-  void set_scan_delay_ms(uint64_t value) { scan_delay_ms_ = value; }
+  uint32_t scan_delay_ms() const { return scan_delay_ms_; }
+  void set_scan_delay_ms(uint32_t value) { scan_delay_ms_ = value; }
 
-  uint64_t scan_batch_size() const { return scan_batch_size_; }
-  void set_scan_batch_size(uint64_t value) { scan_batch_size_ = value; }
+  uint32_t scan_batch_size() const { return scan_batch_size_; }
+  void set_scan_batch_size(uint32_t value) { scan_batch_size_ = value; }
 
   bool do_not_scan() const { return do_not_scan_; }
   void set_do_not_scan(bool value) { do_not_scan_ = value; }
@@ -127,9 +127,9 @@ class PERFETTO_EXPORT InodeFileConfig {
   }
 
  private:
-  uint64_t scan_interval_ms_ = {};
-  uint64_t scan_delay_ms_ = {};
-  uint64_t scan_batch_size_ = {};
+  uint32_t scan_interval_ms_ = {};
+  uint32_t scan_delay_ms_ = {};
+  uint32_t scan_batch_size_ = {};
   bool do_not_scan_ = {};
   std::vector<std::string> scan_mount_points_;
   std::vector<MountPointMappingEntry> mount_point_mapping_;
